add 3scale management dashboard to fuse online custom resource
in fuse online 1.5 or syndesis 1.8 the custom resource, syndesis,
has changed to now require a new flag to specify whether services
should be exposed to 3scale or not.

this addresses this change, if the 3scale url is specified as an
environment variable, it will be used in the syndesis cr.